% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{parcel_segstats}
\alias{parcel_segstats}
\title{Extract voxel values from an input volume based on a parcellation atlas and apply a reduction function to each parcel.}
\usage{
parcel_segstats(
  input_vol,
  atlas_space = "MNI152NLin6Asym",
  atlas = "aseg_subcortex",
  func_name = "Functional map",
  parc_stat = mean,
  ignore_background = TRUE,
  background_value = 0,
  interpolation = NULL
)
}
\arguments{
\item{input_vol}{The input 3D or 4D NIfTI image from which to extract voxel values.
Can be an RNifti object or a file path to a NIfTI image.}

\item{atlas_space}{The standard space to use for the corresponding atlas. Options include
\code{'MNI152NLin6Asym'} (the default) and \code{'MNI152NLin2009cAsym'}.}

\item{atlas}{Name(s) of the subcortical atlas/atlases to apply. Default is
\code{'aseg_subcortex'}. If multiple atlases are provided as a list, the function will
iterate over them and concatenate results.}

\item{func_name}{A name for the functional map being summarized, used for labeling purposes
in the output data frame. Default is \code{'Functional map'}.}

\item{parc_stat}{A function like \code{mean}, \code{sd}, etc. that takes an array of values
and returns a single summary statistic (scalar). Default is \code{mean}. Can also be a
list of functions, in which case the output data frame will have one row per parcel per
summary statistic.}

\item{ignore_background}{If TRUE, the background label (as defined by
\code{background_value}) is skipped when extracting parcel values. Default is \code{TRUE}.}

\item{background_value}{Integer label in the parcellation that represents background
(non-parcel) voxels. Default is \code{0}.}

\item{interpolation}{If the input volume and atlas have different affines or spatial
dimensions, this parameter specifies the interpolation method for resampling the atlas to
match the input volume. Options include \code{'nearest'}, \code{'linear'}, and
\code{'cubic'}. If NULL (default), no resampling is performed and an error will be raised
if affines or dimensions do not match.}
}
\value{
A data frame with one row per parcel per summary statistic, with columns:
\code{stat}, \code{value}, \code{Atlas}, \code{Functional_Map}, \code{region},
\code{Hemisphere}, \code{Region_Index}.
}
\description{
Extract voxel values from an input volume based on a parcellation atlas and apply a reduction function to each parcel.
}
\note{
Users should ensure that the input volume is in the same standard space as the atlas
specified by \code{atlas_space} to avoid issues with affine and spatial dimension
mismatches. If resampling is necessary, users must specify an interpolation method.
}
