% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_individual_regions}
\alias{plot_individual_regions}
\title{Helper function to plot each brain region individually to ensure proper plotting order}
\usage{
plot_individual_regions(
  atlas_SVG_df_labeled,
  line_color = "black",
  line_thickness = 0.5,
  cmap = "plasma",
  discrete = TRUE,
  NA_fill = "#cccccc",
  vmin = NULL,
  vmax = NULL,
  midpoint = NULL
)
}
\arguments{
\item{atlas_SVG_df_labeled}{Dataframe containing individual path nodes from SVG plus atlas information}

\item{line_color}{Color for lines outlining each region (default 'black')}

\item{line_thickness}{Thickness of lines outlining each region (default 0.5)}

\item{cmap}{Colormap to fill in regions (can be character name of palette, vector of colors, or a colormap function)}

\item{discrete}{Whether to plot with discrete data or not (default=T)}

\item{NA_fill}{Color to fill in missing data for regions (string/hex code accepted)}

\item{vmin}{Lower bound for color plotting, useful for plotting multiple versions from the same data}

\item{vmax}{Upper bound for color plotting, useful for plotting multiple versions from the same data}

\item{midpoint}{Value to center colors at between vmin and vmax}
}
\value{
A ggplot2 object containing all of the brain regions for the given atlas
}
\description{
This function accepts a dataframe combining SVG elements and atlas information and plots each region individually
according to the atlas-specific plotting order.
}
