% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{prep_data}
\alias{prep_data}
\title{Helper function to prepare the brain data for plotting}
\usage{
prep_data(atlas, hemisphere, subcortex_data = NULL, value_column = "value")
}
\arguments{
\item{atlas}{Name of the atlas to plot (e.g., 'aseg')}

\item{hemisphere}{Hemisphere to plot ('L', 'R', or 'both')}

\item{subcortex_data}{Empirical data to plot (optional)}

\item{value_column}{If empirical data is passed, which column to plot values from}
}
\value{
A dataframe containing all of the SVG path information together with the atlas data
}
\description{
This function reads in the corresponding SVG and atlas ordering file for the
user-specified atlas and hemisphere combination. The SVG is read in as both a
dataframe and an XML file to enable linking between paths and Inkscape metadata.
}
