% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{resolve_fill_scale}
\alias{resolve_fill_scale}
\title{Helper function to select appropriate color palette and fill type for data}
\usage{
resolve_fill_scale(
  cmap = "viridis",
  discrete = TRUE,
  region_order = NULL,
  NA_fill = "#cccccc",
  vmin = NULL,
  vmax = NULL,
  midpoint = NULL
)
}
\arguments{
\item{cmap}{Palette name in the viridis family, a colour vector (for discrete data), or a palette function (for continuous data)}

\item{discrete}{Whether the data is discrete (e.g., names of regions) or not (e.g., continuous empirical data)}

\item{region_order}{Order in which regions should be filled and plotted}

\item{NA_fill}{Color to fill in missing data for regions (string/hex code accepted)}

\item{vmin}{Lower bound for color plotting, useful for plotting multiple versions from the same data}

\item{vmax}{Upper bound for color plotting, useful for plotting multiple versions from the same data}

\item{midpoint}{Value to center colors at between vmin and vmax}
}
\value{
A fill function that can be applied to discrete or continuous data as appropriate
}
\description{
This function reads in the corresponding SVG and atlas ordering file for the
user-specified atlas and hemisphere combination. The SVG is read in as both a
dataframe and an XML file to enable linking between paths and Inkscape metadata.
}
