% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_atlas_regions}
\alias{get_atlas_regions}
\title{Print the names of regions in a given subcortical atlas}
\usage{
get_atlas_regions(atlas_name)
}
\arguments{
\item{atlas_name}{Name of the atlas to get regions frmo (e.g., 'aseg')}
}
\value{
A list of regions in the specified atlas, ordered by segmentation index
}
\description{
This function takes in the name of a given subcortical/cerebellar atlas (e.g., 'aseg')
and returns the list of regions in that atlas, ordered by segmentation index
}
