% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_subcortical_data}
\alias{plot_subcortical_data}
\title{Main function to plot data in a user-selected subcortical/cerebellar atlas}
\usage{
plot_subcortical_data(
  subcortex_data = NULL,
  atlas = "aseg",
  hemisphere = "L",
  value_column = "value",
  line_color = "black",
  line_thickness = 0.5,
  cmap = "viridis",
  NA_fill = "#cccccc",
  vmin = NULL,
  vmax = NULL,
  midpoint = NULL,
  show_legend = TRUE,
  fill_title = "values",
  fontsize = 12
)
}
\arguments{
\item{subcortex_data}{Empirical data to plot (optional)}

\item{atlas}{Name of atlas to use (default='aseg')}

\item{hemisphere}{Name of hemisphere(s) to use (default='L')}

\item{value_column}{If empirical data is passed, which column to plot values from}

\item{line_color}{Color for lines outlining each region (default='black')}

\item{line_thickness}{Thickness of lines outlining each region (default=0.5)}

\item{cmap}{Colormap to fill in regions (can be character name of palette, vector of colors, or a colormap function)}

\item{NA_fill}{Color to fill in missing data for regions (string/hex code accepted; default="#cccccc")}

\item{vmin}{Lower bound for color plotting, useful for plotting multiple versions from the same data}

\item{vmax}{Upper bound for color plotting, useful for plotting multiple versions from the same data}

\item{midpoint}{Value to center colors at between vmin and vmax}

\item{show_legend}{Whether to plot legend at the bottom (default=TRUE)}

\item{fill_title}{Title to use for legend, if requested (default='values')}

\item{font_size}{Font size for legend (default=12)}
}
\value{
A ggplot2 object containing all of the brain regions for the given atlas
}
\description{
This function accepts a dataframe combining SVG elements and atlas information and plots each region individually
according to the atlas-specific plotting order.
}
